+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `org-agenda` buffers are ignored by `partial-recall`.
 - Prelude package was renamed from `wal-prelude` to just `wal` after
   linting.
+- All library files are now tangled using `wal--tangle-target`.
 
 ### Fixed